Network stack: add IPv6 support
Dual-stack operation (i.e. using an AF_INET6 socket to listen for
IPv4 and IPv6 packets) is supported, for both TCP and UDP servers,
and is enabled by default; it can be disabled via the IPV6_V6ONLY
socket option.

The sin_zero field is being added to struct sockaddr_in to make it
standard-compliant, so that sizeof(struct sockaddr_in) can be used
to set the address length in addrport_to_sockaddr(), which is
cleaner than using sizeof(struct sockaddr).

Closes #1179.